fix-#4: added more testcase,filenames are changed according to the guildlines,modified test cases to handel default HTML validation error message for email input,changes the string constants into images.ts
