+++ frontend/src/__tests__/mocks/node-mock.ts
@@ -1,5 +1,5 @@
 // src/mocks/node.js
 import { setupServer } from 'msw/node'
-import { handlers } from './handelers'
+import { handlers } from './handeler-mock'
  
 export const server = setupServer(...handlers)
