--- frontend/src/constants/Stringconstant.ts
@@ -1,13 +1,0 @@
-export const USERNAME_PLACEHOLDER = "Username"
-export const EMAILINPUT_PLACEHOLDER = "Email"
-export const PASSWORDINPUT_PLACEHOLDER = "Password"
-export const CONFIRMPASSWORD_PLACEHOLDER = "Confirm Password"
-export const SIGNUPBUTTON_TEXT = "Sign Up"
-
-export const INVALID_USERNAME_ERRORMESSAGE = "Username must be at least 5 characters long"
-export const INVALID_EMAIL_ERRORMESSAGE = "Invalid email address"
-export const INVALID_PWD_ERRORMESSAGE = "Password must be at least 8 characters long"
-export const INVALID_CONFIRMPWD_ERRORMESSAGE = "Passwords do not match"
-
-
-
