+++ frontend/src/constants/images.ts
@@ -7,3 +7,18 @@ export const imageUrls: string[] = [
   'https://i.ibb.co/KLwfZzG/Maldives-1024x767.webp',
   'https://i.ibb.co/qxFMj1H/sunset-horizon-clean-sky-nature.webp',
 ];
+export const USERNAME_EMPTY_ERRORMESSAGE = "Username is required"
+export const EMAIL_EMPTY_ERRORMESSAGE = "Email is required"
+export const PASSWORD_EMPTY_ERRORMESSAGE = "Password is required"
+export const CONFIRMPASSWORD_EMPTY_ERRORMESSAGE = "Confirm Password is required"
+
+export const USERNAME_PLACEHOLDER = "Username"
+export const EMAILINPUT_PLACEHOLDER = "Email"
+export const PASSWORDINPUT_PLACEHOLDER = "Password"
+export const CONFIRMPASSWORD_PLACEHOLDER = "Confirm Password"
+export const SIGNUPBUTTON_TEXT = "Sign Up"
+
+export const INVALID_USERNAME_ERRORMESSAGE = "Username must be at least 5 characters long"
+export const INVALID_EMAIL_ERRORMESSAGE = "Invalid email address"
+export const INVALID_PWD_ERRORMESSAGE = "Password must be at least 8 characters long"
+export const INVALID_CONFIRMPWD_ERRORMESSAGE = "Passwords do not match"
