+++ frontend/src/lib/types.ts
@@ -1,4 +1,4 @@
-import { INVALID_CONFIRMPWD_ERRORMESSAGE, INVALID_EMAIL_ERRORMESSAGE, INVALID_PWD_ERRORMESSAGE, INVALID_USERNAME_ERRORMESSAGE } from '@/constants/Stringconstant';
+import { INVALID_CONFIRMPWD_ERRORMESSAGE, INVALID_EMAIL_ERRORMESSAGE, INVALID_PWD_ERRORMESSAGE, INVALID_USERNAME_ERRORMESSAGE } from '@/constants/images';
 import { z } from 'zod';
 
 const emailRegex = /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i;
